Modified for debugging
Testing on the Stm32f03x dev board and stores data to buffers. Further testing will be done with the PowerBoard.
One key error is if the powerboard RX a msg longer then 8 chars, it will read the first 8 and discard the rest. I want the powerboard to disregard the msg completly and return a TX to the user informing them on the TX rules.
ALL COMMS TO PWRBRD NEED TO BE WITHIN 8 CHARS